python_scripts/requirements.txt
face_recognition
dlib
numpy
pymongo
python-dotenv